moved arithmetic_multiple_lines as IT
